+++ frontend/libs/widgets/src/index.ts
@@ -23,6 +23,7 @@ export * from "./settings";
 export * from "./slider-purposes";
 export * from "./telemetry";
 export * from "./topic-suggestions";
+export * from "./urdf-limits";
 export * from "./values";
 export * from "./widget-catalog";
 export * from "./widget-definition";